bckend/requirements.txt
asgiref==3.7.2
certifi==2023.7.22
charset-normalizer==3.2.0
distlib==0.3.7
Django==4.2.5
django-cors-headers==4.2.0
djangorestframework==3.14.0
filelock==3.12.3
idna==3.4
pipenv==2023.9.8
platformdirs==3.10.0
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
sqlparse==0.4.4
urllib3==2.0.4
virtualenv==20.24.5
